refactor: rename modules to gridder and taper
